docs: reorganize the feature list to reflect the refactoring items
