+++ README.md
@@ -1,52 +1,76 @@
 # java-convenience-store-precourse
 ## 기능 목록
-### 입출력
 
-#### 입력
+### 입출력
+> 입력과 출력 관련
+
+#### 입력: StoreInput
 - [x] 구매할 상품과 수량을 입력 받는다. 상품명, 수량은 하이픈(-)으로, 개별 상품은 대괄호([])로 묶어 쉼표(,)로 구분한다.
 - [x] 프로모션 적용이 가능한 상품에 대해 고객이 해당 수량보다 적게 가져온 경우, 그 수량만큼 추가 여부를 입력받는다.
 - [x] 프로모션 재고가 부족하여 일부 수량을 프로모션 혜택 없이 결제해야 하는 경우, 일부 수량에 대해 정가로 결제할지 여부를 입력받는다.
 - [x] 멤버십 할인 적용 여부를 입력 받는다.
 - [x] 추가 구매 여부를 입력 받는다.
 
-#### 출력
+#### 출력: StoreOutput
 - [x] 환영 인사와 함께 상품명, 가격, 프로모션 이름, 재고를 안내한다.
 - [x] 프로모션 적용이 가능한 상품에 대해 고객이 해당 수량만큼 가져오지 않았을 경우, 혜택에 대한 안내 메시지를 출력한다.
 - [x] 프로모션 재고가 부족하여 일부 수량을 프로모션 혜택 없이 결제해야 하는 경우, 일부 수량에 대해 정가로 결제할지 여부에 대한 안내 메시지를 출력한다.
 - [x] 멤버십 할인 적용 여부를 확인하기 위해 안내 문구를 출력한다.
 
-#### Reader
-- [x] 상품 목록과 행사 목록을 파일 입출력을 통해 불러온다.
+#### 특수 입력: YesNoOption
+> StoreInput을 주입받는다.
+- [x] StoreInput으로 입력을 받고, Y 혹은 N만 입력 가능하게 검증한다.
+- [x] 반환값으로 true, 혹은 false로 반환한다.
 
-### FileReader
+#### Reader
+> 정보를 읽어오는 역할을 가진 인터페이스
+
+#### FileReader implements Reader
+- [x] 상품 목록과 행사 목록을 파일 입출력을 통해 불러온다.
 - [x] 파일을 읽어온다.
 
-#### ConvenienceStoreParser
+<br/>
+
+### 변환
+#### ItemParser
+> 상품을 파싱하는 역할을 가진 인터페이스
+
+#### PromotionParser
+> 행사를 파싱하는 역할을 가진 인터페이스
+
+#### Parser extends ItemParser, PromotionParser
+> 각종 파싱 인테페이스를 상속받아 사용하는 인터페이스
+
+#### ConvenienceStoreParser implements Parser
+> 이 편의점 프로그램에서 사용하는 파서를 구현한 클래스
 - [x] 문자열을 기준으로 promotions에 맞는 데이터 타입으로 파싱한다.
 - [x] 문자열을 기준으로 items에 맞는 데이터 타입으로 파싱한다.
 
-### 상태와 동작
+<br/>
 
-#### 재고 하나: Item
+### 재고 관리
+#### 상품: Item
 > 이름, 가격, 프로모션 소유
 
 - [x] 프로모션의 항목에 따라 현재 재고가 얼마나 남았는지 반환한다.
 - [x] 필요한 수량에 맞춰 부족한 수량을 반환한다.
 - [x] 자신의 개수 상태를 변경한다.
 
+#### InventoryFactory
+- [x] Reader에서 읽어온 값을 Parser로 파싱 후 재고를 만들고, Inventory를 생성한다.
+- [x] FileReader, ConvenienceStoreParser 주입을 받는다.
+
 #### 총 재고: Inventory
-> 각 재고 아이템들 소유
+> 각 상품들 소유
 - [x] 재고를 고려하여 결제가 가능한지 알려준다.
 - [x] 재고를 차감할 수 있다.
-- [x] 프로모션 재고가 다 떨어졌을 때에는 일반 재고를 차감한다.(애매하게 1개 남았을 때도)
-- [x] 프로모션 재고 판매 가능 개수를 반환한다.
 - [x] 한 제품의 모든 재고 개수를 반환한다.
 - [x] 만약 재고가 0개라면 `재고 없음`을 반환한다.
 - [x] 현재 재고의 상태를 문자열로 반환한다.
-
-#### 영수증: Receipt
-> 영수증 정보 내역 소유
-- [x] 영수증의 내역을 저장하고, 문자열 값으로 변환한다.
+- [x] 재고 정보에 관해 요청한 정보를 반환한다.
+- [x] 재고를 추가하거나 뺄 수 있다.
+- [x] 현재 재고 정보를 문자열로 반환한다.
+- [x] 재고 상황에 맞춰 프로모션 아이템들과 아닌 아이템들을 조정한다.
 
 
 #### 프로모션: Promotion
@@ -54,70 +78,73 @@
 - [x] 프로모션 기간이 맞는지 확인한다.
 - [x] 해당 프로모션이 맞는지 확인한다.
 
-#### 편의점: ConvenienceStore implements Store
-> 총 재고, 영수증 소유
+<br/>
+
+### 주문 관리
+> 상품을 스캔하고, 결제정보를 반환
+
+#### 포스기: PosMachine
+> **재고**, **스캐너**(바코드 스캐너같은 역할), 유저가 요청한 구매목록인 **카트**, 카트에서 스캔된 상품들인 **스캔상품정보**가 존재
+- [x] 스캐너로 카트의 제품들을 스캔하고, 스캔한상품정보를 추가한다.
 - [x] 구매할 상품의 프로모션을 확인하고 프로모션만큼 해당 수량이 부족한지 판별한다.
 - [x] 구매할 상품의 프로모션을 확인하고, 프로모션 재고가 부족한지 판별한다.
-- [x] 프로모션 할인을 진행하고, 영수증에 내용을 추가한다.
-- [x] 멤버쉽할인 여부에 따라 멤버쉽 할인을 한다.
-- [x] 구매 상품 내역, 증정 상품 내역, 금액 정보가 담긴 영수증을 발행한다.
-- [x] 추가 구매 여부를 확인한다.
-- [x] 인벤토리의 변경 내역을 반영한다.
-
-### 동작
-
-#### InventoryFactory
-- [x] Reader에서 읽어온 값을 Parser로 파싱 후 재고를 만들고, Inventory를 생성한다.
-- [x] FileReader, ConvenienceStoreParser 주입을 받는다.
-
-#### Discount(interface) extends PromotionDiscount, MembershipDiscount
-- [x] 각 할인을 상속받는다.
-
-#### PromotionDiscount(interface)
-> 프로모션 할인
-- [x] 프로모션으로 제공되는 물품의 개수를 계산한다.
-
-#### MembershipDiscount(interface)
-> 멤버십 할인
-- [x] 상품을 할인 퍼센트만큼 할인한 금액을 계산한다.
-- [x] 할인 퍼센트는 주입받는다.
-
-#### ConvenienceStoreDiscount implements Discount
-- [x] 각 할인을 구현한다.
-
-
-## 리팩토링 목록
-- 크게 재고 시스템, 포스(분류) 시스템, 결제시스템으로 분리한다.                                                                           
-- 각 객체가 생성될 때, 검증하고 객체 생성시에 검증 (이번에는 처음에 객체를 생성하는 것이 아닌, 입력값 이후의 객체 생성이 존재하므로 한번에 생성해서 담아두기는 힘들 것 같음.)
-
-### Inventory
-> 재고 정보
-- [x] 재고 정보에 관해 요청한 정보를 반환한다.
-- [x] 재고를 추가하거나 뺄 수 있다.
-- [x] 현재 재고 정보를 문자열로 반환한다.
-- [x] 재고 상황에 맞춰 프로모션 아이템들과 아닌 아이템들을 조정한다.
-
-### OrderInfo
-- 주문 정보를 가지고 있다.
-- 프로모션과 일반 상품으로 나누어져 있다.
-
-### PosMachine
-> **재고**, **스캐너**(바코드 스캐너같은 역할), 유저가 요청한 구매목록인 **카트**, 카트에서 스캔된 상품들인 **스캔상품정보**가 존재 
-- [x] 스캐너로 카트의 제품들을 스캔하고, 스캔한상품정보를 추가한다.
 - [x] 스캔한 상품정보와 재고를 바탕으로 사용자에게 정보를 받고, 반영한 주문정보를 반환한다.
+- [x] 재고의 변경 내역을 반영한다.
 
-### Payment
+#### OrderInfo
+> 주문 정보를 가지고 있다. 프로모션과 일반 상품으로 나누어져 있다.
+
+<br/>
+
+### 결제 관리
+
+#### 결제: Payment
 > 영수증, 할인
 - [x] 프로모션 기간을 확인하여 할인에 반영한다.
 - [x] 주문정보를 바탕으로 할인한다.
 - [x] 결제 정보를 영수증에 추가하여 작성한다.
+
+#### 영수증: Receipt
+> 영수증 정보 내역 소유
+- [x] 영수증의 내역을 저장하고, 문자열 값으로 변환한다.
+- [x] 멤버쉽할인 여부에 따라 멤버쉽 할인을 한다.
+- [x] 프로모션 할인을 진행하고, 영수증에 내용을 추가한다.
+- [x] 구매 상품 내역, 증정 상품 내역, 금액 정보가 담긴 영수증을 발행한다.
+
+#### PromotionDiscount
+> 프로모션 할인을 진행하는 인터페이스
+- [x] 프로모션으로 제공되는 물품의 개수를 계산한다.
+
+#### MembershipDiscount
+> 멤버십 할인을 진행하는 인터페이스
+- [x] 상품을 할인 퍼센트만큼 할인한 금액을 계산한다.
+- [x] 할인 퍼센트는 주입받는다.
+
+#### Discount extends PromotionDiscount, MembershipDiscount
+> 각 할인을 상속받는 인터페이스
+- 각 할인금액을 계산하여 반환한다.
+
+#### ConvenienceStoreDiscount implements Discount
+> 각 할인을 구현하는 클래스
+- [x] 각 할인을 구현한다.
+
+<br/>
+<br/>
+
+## 리팩토링 목록
+- [x] 크게 재고 시스템, 포스(분류) 시스템, 결제시스템으로 분리한다.(완료. 기능목록에 반영됨)                                                                           
+- [x] 각 객체가 생성될 때, 검증하고 객체 생성시에 검증 (이번에는 처음에 객체를 생성하는 것이 아닌, 입력값 이후의 객체 생성이 존재하므로 한번에 생성해서 담아두기는 힘들 것 같음.)
+- [ ] 추가 구매 여부를 확인한다.
+- [ ] writer 로직 변경. 현재 resource 디렉토리 내에서 변경하는 것이 아닌 새로 파일을 만듬
+
+### Inventory
+> 재고 정보
 
 ### 전역
 - [x] 객체 생성을 관리하는 컨테이너 구현(저번 미션 코드)
 - [x] 입력값으로 Y, N을 판별하여 boolean 값으로 반환하는 클래스 구현
 
 ### 수정사항
-- [ ] writer 로직 변경. 현재 resource 디렉토리 내에서 변경하는 것이 아닌 새로 파일을 만듬
 - [x] 프로모션 기간 확인 필요.
 
 ### 다시 설계한 대략적 흐름도
