+++ src/test/java/store/domain/ItemTest.java
@@ -3,7 +3,7 @@ package store.domain;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import store.Promotion;
+import store.stock.Promotion;
 import store.stock.Item;
