+++ src/test/java/store/domain/PromotionTest.java
@@ -2,7 +2,7 @@ package store.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import store.Promotion;
+import store.stock.Promotion;
 
 import static org.junit.jupiter.api.Assertions.*;
